feat/JiraNC-42/TwoColumnsReverseColumnUpgrade
Improvement to the TwoColumns element. This improvement add a Configurations tab and a boolean option to allow users to set up if they need the columns to reverse when viewing from a mobile device.

Also moved the width of both columns to this new configs tab to make the feature more usable.